refactor: merge all random commands
